# src/main/resources/oapi/openapi.yaml
openapi: 3.0.3
info:
  title: GMMS API
  version: 0.0.1

servers:
  - url: http://localhost:3000/api/v1
  - url: https://gmms.mkubik.net/api/v1

paths:

  /gyms:
    post:
      tags:
        - Gym
      operationId: createGym
      summary: Create a new gym
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGymRequest'
      responses:
        '201':
          description: Gym created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGymResponse'
        '400':
          description: Validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Gym with this name already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

    get:
      tags:
        - Gym
      operationId: listGyms
      summary: List all gyms
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: List of gyms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListGymsResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

  /gyms/revenue:
    get:
      tags:
        - Report
      operationId: getRevenueReport
      summary: Get total monthly revenue per gym grouped by currency
      responses:
        '200':
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueReportResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

  /gyms/{gymId}/plans:
    post:
      tags:
        - MembershipPlan
      operationId: createMembershipPlan
      summary: Create a new membership plan for a given gym
      parameters:
        - $ref: '#/components/parameters/GymId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembershipPlanRequest'
      responses:
        '201':
          description: Membership plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMembershipPlanResponse'
        '400':
          description: Validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Gym not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

    get:
      tags:
        - MembershipPlan
      operationId: listMembershipPlans
      summary: List all membership plans for a given gym
      parameters:
        - $ref: '#/components/parameters/GymId'
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: List of membership plans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMembershipPlansResponse'
        '404':
          description: Gym not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

  /plans/{planId}/members:
    post:
      tags:
        - Member
      operationId: registerMember
      summary: Register a new member to a given membership plan
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterMemberRequest'
      responses:
        '201':
          description: Member registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterMemberResponse'
        '400':
          description: Validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '404':
          description: Membership plan not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Membership plan has reached maximum capacity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

  /members:
    get:
      tags:
        - Member
      operationId: listMembers
      summary: List all members including plan name, gym name and status
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMembersResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

  /members/{memberId}/cancel:
    patch:
      tags:
        - Member
      operationId: cancelMembership
      summary: Cancel a membership
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: Membership cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelMembershipResponse'
        '404':
          description: Member not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Membership is already cancelled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'

components:

  parameters:
    GymId:
      name: gymId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    PlanId:
      name: planId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    PageNumber:
      name: page
      in: query
      required: false
      description: Zero-based page index (defaults to 0)
      schema:
        type: integer
        minimum: 0

    PageSize:
      name: size
      in: query
      required: false
      description: Number of items per page (defaults to 20, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
#        default: 20 //TODO: issue here with spring doc

  schemas:

    # -------------------------
    # Enums
    # -------------------------

    MembershipStatus:
      type: string
      enum: [ ACTIVE, CANCELLED ]

    # -------------------------
    # Gym
    # -------------------------

    CreateGymRequest:
      type: object
      required:
        - name
        - address
        - phoneNumber
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "FitLife Center"
        address:
          type: string
          minLength: 1
          maxLength: 255
          example: "ul. Przykładowa 1, Warszawa"
        phoneNumber:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^\+?[0-9 \-().]{7,50}$'
          example: "+48123456789"

    CreateGymResponse:
      type: object
      required:
        - id
        - name
        - address
        - phoneNumber
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        phoneNumber:
          type: string

    GymEntry:
      type: object
      required:
        - id
        - name
        - address
        - phoneNumber
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        phoneNumber:
          type: string

    ListGymsResponse:
      type: object
      properties:
        gyms:
          type: array
          items:
            $ref: '#/components/schemas/GymEntry'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # -------------------------
    # Membership Plan
    # -------------------------

    CreateMembershipPlanRequest:
      type: object
      required:
        - name
        - planTypeCode
        - price
        - currency
        - durationMonths
        - maxMembers
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Morning BASIC"
        planTypeCode:
          # not an enum - plan_type is a lookup table and can be extended at runtime
          type: string
          minLength: 1
          maxLength: 50
          example: "BASIC"
        price:
          type: number
          format: decimal
          exclusiveMinimum: 0
          example: 49.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: "EUR"
        durationMonths:
          type: integer
          minimum: 1
          example: 3
        maxMembers:
          type: integer
          minimum: 1
          example: 20

    CreateMembershipPlanResponse:
      type: object
      required:
        - id
        - name
        - planTypeCode
        - price
        - currency
        - durationMonths
        - maxMembers
        - activeMembers
      properties:
        id:
          type: string
          format: uuid
#         gymId:
#           type: string
#           format: uuid
#        gymName:
#          type: string
#          example: "FitLife Center"
        name:
          type: string
          example: "Morning BASIC"
        planTypeCode:
          type: string
          example: "BASIC"
        price:
          type: number
          format: decimal
          example: 49.99
        currency:
          type: string
          example: "EUR"
        durationMonths:
          type: integer
          example: 3
        maxMembers:
          type: integer
          example: 20
        activeMembers:
          type: integer
          example: 0

    MembershipPlanEntry:
      type: object
      required:
        - id
        - name
        - planTypeCode
        - price
        - currency
        - durationMonths
        - maxMembers
        - activeMembers
      properties:
        id:
          type: string
          format: uuid
#         gymId:
#           type: string
#           format: uuid
#        gymName:
#          type: string
#          example: "FitLife Center"
        name:
          type: string
          example: "Morning BASIC"
        planTypeCode:
          type: string
          example: "BASIC"
        price:
          type: number
          format: decimal
          example: 49.99
        currency:
          type: string
          example: "EUR"
        durationMonths:
          type: integer
          example: 3
        maxMembers:
          type: integer
          example: 20
        activeMembers:
          type: integer
          example: 12

    ListMembershipPlansResponse:
      type: object
      properties:
        plans:
          type: array
          items:
            $ref: '#/components/schemas/MembershipPlanEntry'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # -------------------------
    # Member
    # -------------------------

    RegisterMemberRequest:
      type: object
      required:
        - fullName
        - email
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 255
          example: "Jan Kowalski"
        email:
          type: string
          format: email
          maxLength: 255
          example: "jan@kowalski.pl"

    RegisterMemberResponse:
      required:
        - id
        - fullName
        - email
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
          example: "Jan Kowalski"
        email:
          type: string
          format: email
          example: "jan@kowalski.pl"
        startDate:
          type: string
          format: date
          example: "2026-05-01"
        status:
          $ref: '#/components/schemas/MembershipStatus'
        # planId:
        #   type: string
        #   format: uuid
        planName:
          type: string
          example: "Morning BASIC"
#        gymName:
#          type: string
#          example: "FitLife Center"

    MemberEntry:
      required:
        - id
        - fullName
        - email
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
          example: "Jan Kowalski"
        email:
          type: string
          format: email
          example: "jan@kowalski.pl"
        startDate:
          type: string
          format: date
          example: "2026-05-01"
        status:
          $ref: '#/components/schemas/MembershipStatus'
        # planId:
        #   type: string
        #   format: uuid
        planName:
          type: string
          example: "Morning BASIC"
        gymName:
          type: string
          example: "FitLife Center"

    ListMembersResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberEntry'

    CancelMembershipResponse:
      required:
        - id
        - fullName
        - status
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
          example: "Jan Kowalski"
        status:
          $ref: '#/components/schemas/MembershipStatus'

    # -------------------------
    # Revenue Report
    # -------------------------

    RevenueReportEntry:
      type: object
      properties:
        gymName:
          type: string
          example: "FitLife Center"
        amount:
          type: number
          format: decimal
          example: 2048.00
        currency:
          type: string
          example: "EUR"

    RevenueReportResponse:
      type: object
      properties:
        report:
          type: array
          items:
            $ref: '#/components/schemas/RevenueReportEntry'

    # -------------------------
    # Pagination
    # -------------------------

    PageMetadata:
      type: object
      properties:
        number:
          type: integer
          description: Current page number (zero-based)
          example: 0
        size:
          type: integer
          description: Number of items per page
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of items
          example: 42
        totalPages:
          type: integer
          description: Total number of pages
          example: 3

    # -------------------------
    # Problem Detail (RFC 7807)
    # -------------------------

    ProblemDetailResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
          default: "about:blank"
        title:
          type: string
          nullable: true
        status:
          type: integer
        detail:
          type: string
          nullable: true
        instance:
          type: string
          format: uri
          nullable: true
      additionalProperties: true